+++ server/src/main/java/de/njsm/stocks/server/v2/db/FailSafeDatabaseHandler.java
@@ -55,6 +55,11 @@ public class FailSafeDatabaseHandler implements FallibleOperationWrapper<DSLCont
     }
 
     public StatusCode setReadOnly() {
+        try {
+            connectionFactory.initConnection();
+        } catch (SQLException e) {
+            return StatusCode.DATABASE_UNREACHABLE;
+        }
         return new ConnectionHandler(connectionFactory).setReadOnly();
     }