+++ server/src/test/java/de/njsm/stocks/server/v2/business/EanNumberManagerTest.java
@@ -37,6 +37,7 @@ import java.util.stream.Stream;
 import static de.njsm.stocks.server.v2.web.PrincipalFilterTest.TEST_USER;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.when;
 
 public class EanNumberManagerTest {
 
@@ -62,6 +63,7 @@ public class EanNumberManagerTest {
     public void gettingItemsIsForwarded() {
         AsyncResponse r = Mockito.mock(AsyncResponse.class);
         Mockito.when(backend.get(false, Instant.EPOCH)).thenReturn(Validation.success(Stream.empty()));
+        when(backend.setReadOnly()).thenReturn(StatusCode.SUCCESS);
 
         Validation<StatusCode, Stream<EanNumber>> result = uut.get(r, false, Instant.EPOCH);