+++ server/src/test/java/de/njsm/stocks/server/v2/business/FoodManagerTest.java
@@ -37,6 +37,7 @@ import java.util.stream.Stream;
 import static de.njsm.stocks.server.v2.web.PrincipalFilterTest.TEST_USER;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.when;
 
 public class FoodManagerTest {
 
@@ -71,6 +72,7 @@ public class FoodManagerTest {
     @Test
     public void gettingItemsIsForwarded() {
         AsyncResponse r = Mockito.mock(AsyncResponse.class);
+        when(backend.setReadOnly()).thenReturn(StatusCode.SUCCESS);
         Mockito.when(backend.get(false, Instant.EPOCH)).thenReturn(Validation.success(Stream.empty()));
 
         Validation<StatusCode, Stream<Food>> result = uut.get(r, false, Instant.EPOCH);