+++ server/src/test/java/de/njsm/stocks/server/v2/business/UserManagerTest.java
@@ -40,6 +40,7 @@ import static de.njsm.stocks.server.v2.web.PrincipalFilterTest.TEST_USER;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.when;
 
 public class UserManagerTest {
 
@@ -76,6 +77,7 @@ public class UserManagerTest {
         AsyncResponse r = Mockito.mock(AsyncResponse.class);
         Mockito.when(userDbHandler.get(false, Instant.EPOCH)).thenReturn(Validation.success(Stream.empty()));
         Mockito.when(userDbHandler.commit()).thenReturn(StatusCode.SUCCESS);
+        when(userDbHandler.setReadOnly()).thenReturn(StatusCode.SUCCESS);
 
         Validation<StatusCode, Stream<User>> result = uut.get(r, false, Instant.EPOCH);